Refactor XCM translation logic into separate module
Extract XCM location translation functions from treasury.rs into a
dedicated xcm_translation.rs module for better code organization.

- Move all XCM version handling (V3, V4, V5) to xcm_translation.rs
- Keep treasury.rs focused on treasury-specific migration logic
- Add xcm_translation module to lib.rs
- Maintain same functionality with improved separation of concerns

No functional changes, purely organizational refactoring.
